Removed overview mouseover code, since it doesn't work the way I thought
I had wanted the overview graph to display a different cursor depending
on whether the user was hovering over the highlighted region or not.
Unfortunately, the mouseover callback is only called once when the mouse
enters the graph, not repeatedly when the mouse moves over the graph, so
it can't be used for this purpose.